Fix backend imports and enable CORS support
