+++ backend/app/routes/__init__.py
@@ -1,16 +1,17 @@
 from flask import Blueprint, jsonify, request
-from .models import Usuario
-from . import db
 
-main = Blueprint("main", __name__)
+from ..models import Usuario
+from .. import db
 
-# 🏠 Ruta base
-@main.route("/")
+bp = Blueprint("main", __name__)
+
+
+@bp.route("/")
 def home():
     return jsonify({"message": "Agenda Inteligente API funcionando correctamente ✅"})
 
-# 👥 Obtener todos los usuarios
-@main.route("/usuarios", methods=["GET"])
+
+@bp.route("/usuarios", methods=["GET"])
 def obtener_usuarios():
     usuarios = Usuario.query.all()
     resultado = [
@@ -19,14 +20,14 @@ def obtener_usuarios():
             "nombre": u.nombre,
             "correo": u.correo,
             "zona_horaria": u.zona_horaria,
-            "creado_en": u.creado_en
+            "creado_en": u.creado_en,
         }
         for u in usuarios
     ]
     return jsonify(resultado), 200
 
-# ➕ Crear un nuevo usuario
-@main.route("/usuarios", methods=["POST"])
+
+@bp.route("/usuarios", methods=["POST"])
 def crear_usuario():
     datos = request.get_json()
     if not datos or "nombre" not in datos or "correo" not in datos:
@@ -35,15 +36,23 @@ def crear_usuario():
     nuevo = Usuario(
         nombre=datos["nombre"],
         correo=datos["correo"],
-        zona_horaria=datos.get("zona_horaria", "America/Mexico_City")
+        zona_horaria=datos.get("zona_horaria", "America/Mexico_City"),
     )
     db.session.add(nuevo)
     db.session.commit()
 
-    return jsonify({
-        "id": nuevo.id,
-        "nombre": nuevo.nombre,
-        "correo": nuevo.correo,
-        "zona_horaria": nuevo.zona_horaria,
-        "creado_en": nuevo.creado_en
-    }), 201
+    return jsonify(
+        {
+            "id": nuevo.id,
+            "nombre": nuevo.nombre,
+            "correo": nuevo.correo,
+            "zona_horaria": nuevo.zona_horaria,
+            "creado_en": nuevo.creado_en,
+        }
+    ), 201
+
+
+# Retrocompatibilidad con importaciones antiguas
+main = bp
+
+__all__ = ["bp", "main"]
