+++ backend/requirements.txt
@@ -1,5 +1,6 @@
 Flask==3.0.3
 Flask-SQLAlchemy==3.1.1
+Flask-Cors==4.0.0
 PyMySQL==1.1.1
 python-dotenv==1.0.1
 requests==2.31.0
